Strip repo down to typescript-example only
The original repository contained many packages, each of which was a
self-contained, example intern setup. The purpose of this repo is to
reproduce a specific issue with VS Code, so we require only one of
these packages. Thus, this commit removes all the other
packages and moves the files in the `typescript-example` package to
be at the top level of the repo.